+++ weldx/util/xarray.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from collections.abc import Iterable
-from typing import Any, Dict, List, Tuple, Union
+from typing import TYPE_CHECKING, Any, Dict, List, Tuple, Union
 
 import numpy as np
 import pandas as pd
@@ -16,6 +16,9 @@ from scipy.spatial.transform import Slerp
 
 from weldx.constants import Q_, U_, UNITS_KEY
 from weldx.time import Time, types_time_like, types_timestamp_like
+
+if TYPE_CHECKING:  # pragma: no cover
+    import weldx.types as wxt
 
 __all__ = [
     "WeldxAccessor",
@@ -534,7 +537,7 @@ def xr_check_coords(dax: xr.DataArray, ref: dict) -> bool:
 
 
 def xr_3d_vector(
-    data: np.ndarray,
+    data: wxt.ArrayLike,
     time: types_time_like = None,
     add_dims: List[str] = None,
     add_coords: Dict[str, Any] = None,
@@ -590,7 +593,7 @@ def xr_3d_vector(
     return da.astype(float).weldx.time_ref_restore()
 
 
-def xr_3d_matrix(data: np.ndarray, time: Time = None) -> xr.DataArray:
+def xr_3d_matrix(data: wxt.ArrayLike, time: Time = None) -> xr.DataArray:
     """Create an xarray 3d matrix with correctly named dimensions and coordinates.
 
     Parameters